Ignore box free more selectively
Drop elaboration introduces box_free calls which
are not always allowed by borrowck.
In particular, borrowck can't correlate the value
of drop flags to the initializations status of their
corresponding variable, so any box_free after a drop
flag read will present problems to the borrowck as
its arguments will be flagged as maybe uninitialized.
However, this is already what is happening with Drops
where borrowck trusts drop elaboration to do the right
thing, and we just extend this logic to box_free.