Derive traversable impls for BindingForm
BindingForm has been a no-op traversable ever since it was first added
in cac6126, but it was modified in 0193d1f to (optionally) contain a
Place without its implementations being updated to fold/visit such.

By using the derive macro to implement the visitable traits, we ensure
that all contained types of interest are folded/visited.