std: add predicate pattern support to OsStr
Due to technical limitations adding support for predicate as patterns
on OsStr slices must be done via core::pattern::Predicate wrapper type.
This isn’t ideal but for the time being it’s the best option I’ve came
up with.

The core of the issue (as I understand it) is that FnMut is a foreign
type in std crate where OsStr is defined.

Using predicate as a pattern on OsStr is the final piece which now
allows parsing command line arguments.