Pass in the layout to `codegen_rvalue_operand`/`rvalue_creates_operand`
It turns out that every caller either had it easily available or had calculated it previously but thrown it away.

I started doing this because I wanted to look at the layout in order to do better transmute codegen -- which I left out of this PR -- but hopefully it'll be good in general too.