1.68 added the creation of a `build/host` symlink  that points to the `build/<host_triple>` directory. However, this directory does not always exist.
The directory is normally created when downloading the bootstrap tools,
but if the user manually configures `build.rustc` and `build.cargo` then
the build will fail.

This fixes the issue be creating the directory junction target if it
does not exist. It's not an issue on non-Windows platforms,
because they use symlinks (which are allowed to point to non-existent
entities).