Use `ty::BoundVar` in more places.
There are lots of conversions between integers and `BoundVars`. Some of
these are unavoidable, but by storing bound vars as `BoundVar` rather
than `u32` in a few places (e.g. `BoundRegionKind::BrAnon`,
`BoundTyKind::Anon`) we reduce the number of conversions. It's also good
to store these values with the more informative type.

The commit also impls `AddAssign<usize>` for newtypes, to allow
incrementing.

Some of the Chalk types (e.g. `BoundVar` and `Placeholder`) use `usize`
for bound vars. If they were change then more conversions could be
avoided, but this is difficult because `chalk_ir` is a module outside
the compiler.