Turn pointer_structural_match forward compat lint into normal lint
This commit removes the pointer_structural_match forward compat lint
from the forward compatibility warning mechanism, as there is no concrete
plan for removal of matching on function pointers.

It also turns the lint from allow-by-default to warn-by-default,
as the behaviour might be surprising to users and they should specifically
opt into the dangers.
As the lint is recognized on older compilers, users can easily allow it
after having made the descision on whether they want to keep it in their
code or not.