Allow older LLVM versions to have missing components
This check was introduced by #77280 to ensure that all tests that are
filtered by LLVM component are actually tested in CI. However this
causes issues for new targets (e.g. #101069) where support is only
available on the latest LLVM version.

This PR restricts the tests to only CI jobs that use the latest LLVM
version.