Require Python 3 in bootstrap.py
If x.py was executed with Python 2, it tries to reexecute itself with
Python 3 if that is available. If it isn't, we now raise an error.

This may break people who have Python 3 in a nonstandard location under
a nonstandard name. These people will now be forced to use
`their-python3 x.py` instead of `./x.py`.