Add -Zsplit-thin-lto-unit flag to enable ThinLTO unit splitting
Clang enables the EnableSplitLTOUnit LLVM module flag when compiling
for ThinLTO when:
* -fsplit-lto-unit is specified, or
* -fsanitize=cfi is specified, or
* -fwhole-program-vtables.

When linking Rust and C++ together and using ThinLTO, the setting
for the LLVM module flag "EnableSplitLTOUnit" must match or linking
simply fails.

The ability to enable this LLVM module flag is then required in
order to link Rust objects with C++ objects with ThinLTO for any
project that uses the above flags (like Chromium does).