Safe Transmute: Pass ErrorGuaranteed up to error reporting
This patch updates the `Answer` enum used by Safe Transmute to have a variant
for errors, and includes the `ErrorGuaranteed` type so that we know that an
error was already emitted. Before this change, we would return `Answer::Yes`
for failures, which gives the same end-result, but is confusing because
`Answer::Yes` is supposed to mean that safe transmutation is possible. Now,
it's clear when we have an error, and we can decide what to do with it during
error reporting. Also, we now know for sure that an error was emitted for
`LayoutError` (previously, we just assumed without confirming).