Specify `--pre-release` when publishing vsce nightly
According to https://code.visualstudio.com/api/working-with-extensions/publishing-extension#prerelease-extensions, when publishing pre-release versions, `--pre-release` must also be passed in the publish step.

Currently the behavior with and without the flag seems to be the same, but since it's documented to be required, don't rely on this staying that way.

This flag was previously removed in #13020, so this partially reverts that change.